authentication modified a bit, getMyApartments implemented, tried to link secondary key of facilities to getApartments

+++ server/controller/locations.js
@@ -1,6 +1,7 @@
 const allLocations = require('../data/locations');
 const { getRequestData } = require('../utils/utils');
 const LocationService = new (require('../service/locationService'))();
+const authenticationService = new (require('../controller/authentication'))();
 const notFound = (id) => ({ message: `Location with id ${id} not found` });
 
 class LocationsControler {
@@ -12,6 +13,15 @@ class LocationsControler {
                 location.position.lng <= geoLocation._northEast.lng &&
                 location.position.lat >= geoLocation._southWest.lat && location.position.lng >= geoLocation._southWest.lng
         });
+    }
+
+    async getLocationsOwnedByUser(request,response)
+    {
+        const payload = await authenticationService.getJWTPayload(request);
+        const locations = await LocationService.getLocationsForUser(payload.id);
+
+        response.writeHead(200, '200', { 'Content-Type': 'application/json' });
+        response.end(JSON.stringify(locations));
     }
     // TO BE REWRITTEN
     async getLocation(id, geoLocation) {
@@ -33,14 +43,10 @@ class LocationsControler {
     }
     //////////////////////////////////////////////
     async getAllLocations(_, response) {
-       
         const locations = await LocationService.getAllLocations();
-
         response.writeHead(200, '200', { 'Content-Type': 'application/json' });
         response.end(JSON.stringify(locations));
     }
-
-
 
     async getTheLocation(request, response) {
         console.log("GET THE LOCATION CALLED");
@@ -53,8 +59,11 @@ class LocationsControler {
     }
 
     async addLocation(request, response) {
-        let location_data = await getRequestData(request);
-        let location = await LocationService.createLocation(JSON.parse(location_data));
+        let location_data = JSON.parse(await getRequestData(request));
+       let payload = await authenticationService.getJWTPayload(request);
+
+        location_data.ownerId = parseInt(payload.id);
+        let location = await LocationService.createLocation(location_data);
 
         response.writeHead(200, '200', { 'Content-Type': 'application/json' });
         response.end(JSON.stringify(location));
@@ -63,27 +72,19 @@ class LocationsControler {
    async deleteLocation (request, response){
         const urlData = request.url.split('/');
         const id = urlData[urlData.length - 1];
-    
+
+        await authenticationService.verifyIfAuthorizedToDelete(parseInt(id),request,response);
         let deletedLocation = await LocationService.deleteLocation(id);
     
         response.writeHead(200, '200', { 'Content-Type': 'application/json' });
         response.end(JSON.stringify(deletedLocation));
     }
 
-
-
-    async createLocation(location) {
-        return new Promise.resolve();
-    }
-
     async updateLocation(request, response) {
         const urlData = request.url.split('/');
         const id = urlData[urlData.length - 1];
-
         let location_data = await getRequestData(request);
-
         let updatedLocation = await LocationService.updateLocation(id, JSON.parse(location_data));
-
         response.writeHead(200, '200', { 'Content-Type': 'application/json' });
         response.end(JSON.stringify(updatedLocation));
     }
